requirements.txt
asgiref==3.6.0
click==8.1.3
Django==4.2.1
django-ninja==0.21.0
markdown-it-py==2.2.0
mdurl==0.1.2
psycopg2-binary==2.9.6
pydantic==1.10.7
Pygments==2.15.1
rich==13.3.5
sqlparse==0.4.4
typing_extensions==4.5.0
